1. Added DataTable unit tests. 2. Added MysqlClient unit tests. 3. Added SqliteClient unit tests. 4. Added SqlSelectFilter unit tests. 5. Added Variant unit tests. 6. Fixed bugs.

+++ tests/TimeZoneTest.cpp
@@ -181,7 +181,7 @@ bool testProperty() {
 
     {
         TimeZone test(TimeZone::GMT_A0800);
-        if (test.ticksOffset() != -8L * 3600 * 1000 * 10000) {
+        if (test.ticksOffset() != -(int64_t)8L * 3600 * 1000 * 10000) {
             return false;
         }
     }
